% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{bisection}
\alias{bisection}
\title{[内部関数] 二分法}
\usage{
bisection(f, interval, tol = .Machine$double.eps * 16)
}
\arguments{
\item{f}{解を探索する関数。}

\item{interval}{解を探索する範囲のベクトル。}

\item{tol}{許容する誤差 (デフォルト: .Machine$double.eps * 16)。}
}
\value{
方程式 f = 0 の解。
}
\description{
二分法により方程式を解く。
}
