% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{bisection.sub}
\alias{bisection.sub}
\title{[内部関数] 二分法用サブ関数}
\usage{
bisection.sub(f, interval, tol)
}
\arguments{
\item{f}{解を探索する関数。}

\item{interval}{解を探索する範囲のベクトル。}

\item{tol}{許容する誤差。}
}
\value{
方程式 f = 0 の解。
}
\description{
二分法の処理実体。
}
