% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\docType{class}
\name{CGD-class}
\alias{CGD-class}
\alias{CGD}
\title{連結ガウス分布 (Connected Gaussian Distribution) クラス}
\description{
平均値の等しい正規分布をX軸方向に連結したモデルを表すクラス
}
\details{
連結ガウス分布は、任意のクォンタイルをゼロ誤差で満たす分布を構成できる。
そのため、何かクォンタイルが与えられたときに、
それを高精度に再現する分布が欲しいときなどに使うことができる。
ただし、確率密度関数は一般に不連続であり、歪んだ形になる。

連結ガウス分布を使って滑らかなグラフを得るには、
ランダムサンプルを取得して、そのヒストグラムを描画するとよい。

クォンタイルが4点以下 (確率50％の点が1点、それ以外の点が3点以下) の場合は、
クォンタイルの位置が過度にいびつでなければ、連続で滑らかな確率密度関数を持つ
確率分布を構成することができる。
}
\section{Fields}{

\describe{
\item{\code{mean}}{平均値}

\item{\code{intervals}}{連結区間 (CGDInterval クラスのリスト)}

\item{\code{type1.type}}{接続区間が type 1 の場合の計算方法 (詳細は [set.waypoints()] を参照)}

\item{\code{m.sd}}{計算済みの標準偏差 (クラス外からは直接参照しないこと)}
}}


\seealso{
\link[cgd]{CGDInterval-class}
\link[cgd]{CGD$set.waypoints}
\href{https://github.com/Kimitsuna-Goblin/CGD}{README.md} (GitHub)
}
