% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$adjust.kind.index}
\alias{CGD$adjust.kind.index}
\title{分布の種類のインデックス番号設定}
\usage{
CGD$adjust.kind.index()
}
\value{
分布の種類のインデックス番号 (フィールドにも同じ値が設定される)
}
\description{
分布の種類のインデックス番号を正しく設定する。
通常、このメソッドは、クラスの利用者が直接呼び出す必要はない。
}
