% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$clear}
\alias{CGD$clear}
\title{フィールド初期化}
\usage{
CGD$clear(clear.intervals = TRUE)
}
\arguments{
\item{clear.intervals}{intervals フィールドを初期化するかどうかのフラグ
 (デフォルト: TRUE)。
FALSE にすると、 intervals フィールドはクリアされないが、
mean などの他のフィールドはクリアされてしまうので、
他のフィールドを明示的に再設定しないと、
オブジェクトは使い物にならなくなるので注意。}
}
\value{
なし
}
\description{
コンストラクタを使わずに、 type1.type 以外のフィールドを初期状態にする。
}
\examples{
 a <- CGD$new()
 a$set.waypoints(
     data.frame( p = c( 0.25, 0.5, 0.75 ), q = c( -0.67, 0, 0.53 ) ),
     this.type1.type = 2, continuous = TRUE )
 a
 a$clear()
 a
}
