% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$get.adjusted.mean}
\alias{CGD$get.adjusted.mean}
\title{平均値取得}
\usage{
a$get.adjusted.mean()
}
\value{
平均値 (intervals 未設定時は NaN)
}
\description{
\link[cgd]{CGD} クラスオブジェクトの mean 以外のフィールドの内容から、平均値を取得する。
通常、このメソッドは、クラスの利用者が直接呼び出す必要はない。
平均値を使用するには、 mean フィールドを使えば十分である。
このメソッドは、クラス内部の処理で intervals フィールドを直接書き換えたときに、
mean フィールドの値を再設定するために用意している。
}
