% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$get.adjusted.median}
\alias{CGD$get.adjusted.median}
\title{中央値取得}
\usage{
CGD$get.adjusted.median()
}
\value{
連結ガウス分布の中央値 (intervals 未設定時は NaN)
}
\description{
\link[cgd]{CGD} クラスオブジェクトの intervals フィールドの内容から、
その連結ガウス分布の中央値を取得する。
通常、このメソッドは、クラスの利用者が直接呼び出す必要はない。
中央値を使用するには、 median フィールドを使えば十分である。
このメソッドは、クラス内部の処理で intervals フィールドを直接書き換えたときに、
median フィールドの値を再設定するために用意している。
}
