% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$is.conn.p}
\alias{CGD$is.conn.p}
\title{接続区間判定 (確率)}
\usage{
a <- CGD$new()
a$is.conn.p(p)
}
\arguments{
\item{p}{確率のベクトル}
}
\value{
第1要素 (bool) に、接続区間に入っていれば TRUE、入っていなければ FALSE、
        第2要素 (i.1) に、前の intervals のインデックス
         (bool = FALSE の場合は 独立区間の intervals のインデックス)、
        第3要素 (i.2) に、後の intervals のインデックス (i.1 + 1, bool = FALSE の場合は NaN) が入ったリスト
}
\description{
指定された確率が接続区間に入っているかどうかを調べる
}
