% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$is.ind.p}
\alias{CGD$is.ind.p}
\title{独立区間判定 (確率)}
\usage{
a <- CGD$new()
a$is.ind.p(p)
}
\arguments{
\item{p}{確率のベクトル}
}
\value{
第1要素 (bool) に、独立区間に入っていれば TRUE、入っていなければ FALSE。
第2要素 (i) に、当該 intervals のインデックス (bool = FALSE の場合は NaN) が入ったリスト
}
\description{
指定された確率が独立区間に入っているかどうかを調べる
}
