% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$is.ind.q}
\alias{CGD$is.ind.q}
\title{独立区間判定 (X座標 (クォンタイル) )}
\usage{
a <- CGD$new()
a$is.ind.q(q)
}
\arguments{
\item{q}{X座標 (クォンタイル) のベクトル}
}
\value{
第1要素 (bool) に、独立区間に入っていれば TRUE、入っていなければ FALSE。
第2要素 (i) に、当該 intervals のインデックス (bool = FALSE の場合は NaN) が入ったリスト
}
\description{
指定されたX座標 (クォンタイル) が独立区間に入っているかどうかを調べる
}
