% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$is.uni.mean}
\alias{CGD$is.uni.mean}
\title{uni.mean 判定}
\usage{
a <- CGD$new()
a$is.uni.mean()
}
\value{
構成するすべての正規分布の平均値が等しければ TRUE、そうでなければ FALSE
}
\description{
uni.mean (分布を構成するすべての正規分布の平均値が等しい) かどうかを調べる
}
