% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$is.v.grad}
\alias{CGD$is.v.grad}
\title{v.grad 判定}
\usage{
CGD$is.v.grad()
}
\value{
v.grad = TRUE として構成されていれば TRUE、そうでなければ FALSE
}
\description{
v.grad = TRUE として構成されているかどうかを調べる。
}
\examples{
 a <- CGD$new()
 a$kind  ## Normal Distribution
 a$is.v.grad()   ## FALSE

 a$set.waypoints(
     data.frame( p = c( 0.1, 0.4, 0.6, 0.9 ), q = c( -1.92, -0.20, 0.20, 1.92 ) ),
     this.type1.type = 3, v.grad = TRUE )
 a$kind  ## Median-Differed Sigma-Differed Vertical Gradational Distribution
 a$is.v.grad()   ## TRUE
}
