% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$p}
\alias{CGD$p}
\title{確率取得}
\usage{
CGD$p(q)
}
\arguments{
\item{q}{X座標 (クォンタイル) のベクトル。}
}
\value{
X座標 (クォンタイル) q における確率
}
\description{
X座標 (クォンタイル) を指定して、確率を取得する。
}
\examples{
a <- CGD$new()
a$set.waypoints(
	data.frame( p = c( 0.25, 0.5, 0.75 ), q = c( -0.67, 0, 0.53 ) ),
	this.type1.type = 2, continuous = TRUE )
a$p( c( -0.67, 0, 0.53 ) )
plot( seq( -3, 3, 0.01 ), a$p( seq( -3, 3, 0.01 ) ), type = "l" )
}
