% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$q}
\alias{CGD$q}
\title{X座標 (クォンタイル) 取得}
\usage{
a <- CGD$new()
a$q(prob, tol = .Machine$double.eps * 16)
}
\arguments{
\item{prob}{確率のベクトル}

\item{tol}{許容誤差 (デフォルト: .Machine$double.eps * 16)}
}
\value{
X座標 (クォンタイル)
}
\description{
確率を指定して、X座標 (クォンタイル) を取得する
}
