% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$set.waypoints}
\alias{CGD$set.waypoints}
\title{経路設定}
\usage{
a <- CGD$new()
a$set.waypoints(waypoints, continuous = FALSE, symmetric = FALSE, this.type1.type = NULL)
}
\arguments{
\item{waypoints}{経路の data.frame( q = 経路のX座標 (クォンタイル), p = その点における確率 )
X座標 (クォンタイル) は昇順にソートしておくこと。
平均値は p = 0.5 の点のX座標として与えること (p = 0.5 の点は必須)}

\item{continuous}{独立区間を [0, 0] と [1, 1] の2点にして、
確率密度関数が全区間 (-∞, ∞) で連続になるように分布構成を試みる (デフォルト: FALSE)}

\item{symmetric}{独立区間を [0, 0], [0.5, 0.5], [1, 1] の3点にして、
                            1番目と3番目の確率分布を同一にすることにより、
                            確率密度関数が全区間 (-∞, ∞) で連続で、かつ左右対称になるように試みる  (デフォルト: FALSE)

            continuous または symmetric を TRUE にする場合、経路の構成点は一定の個数でなければならない。
            具体的には、構成点の個数は以下のようにする必要がある (構成点の個数には、確率 0.5 の点の分を含む)。

            ・continuous = TRUE : type1.type = 1 or 2 ⇒ 3点 (type1.type = 3 では無効)

            ・symmetric = TRUE : type1.type = 1 or 2 ⇒ 3点、type1.type = 3 ⇒ 3点 or 4点}

\item{this.type1.type}{フィールドの type1.type に設定する値 (1、2、3 のいずれかを指定すること) (デフォルト: NULL)
                            NULL の場合は type1.type の値を変更しない

            type1.type = 1 は、最も単純な連結である。2つの正規分布の平均が構成可能である。

            type1.type = 2 は、2つの確率密度関数の横方向グラデーションが構成可能である。

            type1.type = 3 は、2つの確率密度関数の縦方向グラデーションが構成可能である。

            type1.type の値によって、接続区間 (β_i, α_{i+1}) が type 1 の場合、
            接続区間の累積分布関数 Ψ_i(x) を以下のように計算する。

            1: Ψ_i(x) = ( α_{i+1} - x ) / ( α_{i+1} - β_i ) * Φ_i(x) +
                         ( x - β_i ) / ( α_{i+1} - β_i ) * Φ_{i+1}(x)

            2: Ψ_i(x) = ( Φ~_i(α_{i+1}) - Φ~_i(x) ) / ( Φ~_i(α_{i+1}) - Φ~_i(β_i) ) * Φ_i(x) +
                         ( Φ~_i(x) - Φ~_i(β_i) ) / ( Φ~_i(α_{i+1}) - Φ~_i(β_i) ) * Φ_{i+1}(x)

            3: Ψ_i(x) = ∫_{-∞}^x { ( 1 - f_i(t) / f_i(μ) ) f_i(t) + f_{i+1}(t)^2 / f_{i+1}(μ) } dt
                       = Φ_i(x) - Φ^*_i(x) / √2 + Φ^*_{i+1}(x) / √2

            ただし、Φ_i, Φ_{i+1} は接続区間の前後の独立区間における累積分布関数。
            Φ~_i(x) = ( Φ_i(x) + Φ_{i+1}(x) ) / 2。
            f_i, f_{i+1} は接続区間の前後の独立区間における確率密度関数。μ は平均値。
            Φ^*_i は正規分布 N(μ, (σ_i / √2)^2 ) の累積分布関数。

            type1.type = 1 では、
            set.waypoints() の引数で continuous または symmetric を TRUE にした場合 (2つの正規分布の平均)、
            独立区間は [0, 0], [1, 1] の2点になり、
            接続区間の累積分布関数は Ψ(x) = ( Φ_i(x) + Φ_{i+1}(x) ) / 2 となる
             (この場合、独立区間を [0, 0], [0.5, 0.5], [1, 1] の3点と見做してもよい)。

            type1.type = 2 では、
            set.waypoints() の引数で continuous を TRUE にした場合 (横方向グラデーション)、
            独立区間は [0, 0], [1, 1] の2点になり、
            接続区間の累積分布関数は上の 2 の式になる。
            set.waypoints() の引数で symmetric を TRUE にした場合、
            独立区間は [0, 0], [0.5, 0.5], [1, 1] の3点になり、
            接続区間 (0, 0.5) の累積分布関数 Ψ_1(x) は上の 2 の式で、
            接続区間 (0.5, 1) の累積分布関数 Ψ_2(x) は Ψ_2(x) = 1 - Ψ_1(2μ - x) となる。

            type1.type = 3 では、
            独立区間は [0, 0], [0.5, 0.5], [1, 1], [0, 0.5], [0.5, 1], [0, 1] の6種類しか取り得ない。
            set.waypoints() の引数の continuous は無効になる。
            set.waypoints() の引数で symmetric を TRUE にした場合 (縦方向グラデーション)、
            独立区間は [0, 0], [0.5, 0.5], [1, 1] の3点になり、
            接続区間 (0, 0.5), (0.5, 1) の累積分布関数はどちらも同じ、上の 3 の式になる。

            type1.type = 2 と type1.type = 3 で symmetric を TRUE にした場合、
            [0, 0] と [1, 1] の2点が1つめの正規分布 N_1 の独立区間、
            [0.5, 0.5] が2つめの正規分布 N_2 の独立区間となる。}
}
\value{
nleqslv() を内部で実行した場合はその結果、それ以外は NULL
}
\description{
累積分布関数の経路 (クォンタイル) を設定する
}
\seealso{
\href{https://github.com/Kimitsuna-Goblin/CGD}{README.md} (GitHub)
}
