% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\docType{class}
\name{CGDInterval-class}
\alias{CGDInterval-class}
\alias{CGDInterval}
\title{連結ガウス分布区間クラス}
\description{
連結ガウス分布クラス (\link[cgd]{CGD-class}) で使用する区間を表すクラス
}
\section{Fields}{

\describe{
\item{\code{mean}}{平均値}

\item{\code{sd}}{標準偏差}

\item{\code{q.ind}}{独立区間におけるX座標 (クォンタイル)}

\item{\code{q.conn.prev}}{前の区間との接続区間の確率内に収まる、この標準偏差のX座標 (クォンタイル)}

\item{\code{q.conn.next}}{次の区間との接続区間の確率内に収まる、この標準偏差のX座標 (クォンタイル)}

\item{\code{p.ind}}{独立区間における確率}

\item{\code{p.conn.prev}}{前の区間との接続区間の確率}

\item{\code{p.conn.next}}{次の区間との接続区間の確率}
}}


\seealso{
\link[cgd]{CGD-class}
}
