% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGDInterval$q.manage}
\alias{CGDInterval$q.manage}
\title{負担区間取得}
\usage{
a <- CGD$new()
a$intervals[[1]]$q.manage()
}
\value{
負担区間のX座標 (クォンタイル)
}
\description{
負担区間 (前の区間との接続区間～次の区間との接続区間) のX座標 (クォンタイル) を取得する
}
