% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{d}
\alias{d}
\alias{\S4method{d}{CGD}}
\title{確率密度関数}
\usage{
\S4method{d}{CGD}(x)
}
\arguments{
\item{x}{X座標のベクトル。}
}
\value{
x における確率密度関数の値
}
\description{
X座標を指定して、確率密度を求める。
}
\examples{
 a <- CGD$new()
 a$set.waypoints(
     data.frame( p = c( 0.25, 0.5, 0.75 ), q = c( -0.67, 0, 0.53 ) ),
     this.type1.type = 2, continuous = TRUE )
 a$d( c( -0.67, 0, 0.53 ) )
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )
}
