% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{dp.t3}
\alias{dp.t3}
\title{[内部関数] 確率密度関数・累積分布関数計算 (type1.type=3)}
\usage{
dp.t3(x, means, sds, f.t3)
}
\arguments{
\item{x}{X座標のベクトル。}

\item{means}{intervals の平均値のベクトル。}

\item{sds}{intervals の標準偏差のベクトル。}

\item{f.t3}{確率密度 / 累積分布用関数ハンドル ( f.t3.d / f.t3.p )。}
}
\value{
確率密度関数または累積分布関数の値のベクトル
}
\description{
type1.type=3 の確率密度関数または累積分布関数の値を得る
}
