% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{gen.t3.intervals}
\alias{gen.t3.intervals}
\title{type1.type=3 の intervals を生成する}
\usage{
gen.t3.intervals(means, sds)
}
\arguments{
\item{means}{平均値のベクトル}

\item{sds}{標準偏差のベクトル}
}
\value{
生成した intervals
}
\description{
type1.type=3 の intervals を生成する
}
