% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{get.d}
\alias{get.d}
\title{確率密度関数の目標値取得}
\usage{
get.d(x, freq, total)
}
\arguments{
\item{x}{X座標のベクトル。
最低 3 個以上の要素を含み、昇順に並べておくこと。
また、値に重複がないようにすること。}

\item{freq}{度数分布のベクトル。要素数は x と同数であること。}

\item{total}{度数の合計。}
}
\value{
座標 x における確率密度関数の目標値のベクトル
}
\description{
度数分布から確率密度関数の目標値を得る。
}
\examples{
 x <- seq( -2, 2, 0.2 )
 freq <- c( 15.164, 22.923, 25.134, 27.631, 37.239, 40.464,
         47.126, 79.469, 109.966, 118.241, 111.333, 78.674,
         46.921, 41.026, 36.975, 27.403, 25.493, 22.838,
         14.992, 11.468, 9.174 )
 total <- sum( freq )
 cgd::get.d( x, freq, total )
 plot( x, cgd::get.d( x, freq, total ) )
}
