% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{get.intervals.with.nls.coef}
\alias{get.intervals.with.nls.coef}
\title{[内部関数] nls の実行結果による intervals 取得}
\usage{
get.intervals.with.nls.coef(
  coefs,
  normal,
  symmetric,
  v.grad,
  uni.sigma,
  uni.mean,
  type1.type
)
}
\arguments{
\item{coefs}{nls の実行結果の coef (名前付き double 型ベクトル)。}

\item{normal}{正規分布で近似するフラグ。}

\item{symmetric}{type1.type = 2, symmetric = TRUE で近似するフラグ。}

\item{v.grad}{type1.type = 3, v.grad = TRUE で近似するフラグ。}

\item{uni.sigma}{uni.sigma = TRUE で近似するフラグ。}

\item{uni.mean}{構成要素の正規分布の平均値をすべて同じ値に揃えるかどうかのフラグ。}

\item{type1.type}{フィールドの type1.type の値。}
}
\value{
intervals のリスト
}
\description{
nls の実行結果をもとに、 intervals を取得する
}
