% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{get.tex.val}
\alias{get.tex.val}
\title{[内部関数] 変数部の TeX 形式取得}
\usage{
get.tex.val(cgd, decimal)
}
\arguments{
\item{cgd}{\link[cgd]{CGD} クラスオブジェクト。}

\item{decimal}{小数点以下桁数。}
}
\value{
平均値と標準偏差の TeX 形式
}
\description{
平均値と標準偏差の TeX 形式を取得する
}
