% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{init.start.list}
\alias{init.start.list}
\title{nls.freq.all の start 用リスト初期化}
\usage{
init.start.list()
}
\value{
length( cgd:::kinds ) - 1 個の NULL が入ったリスト
}
\description{
要素数が length( cgd:::kinds ) - 1 と等しい NULL のリストを得る。
\link[cgd]{nls.freq.all} の初期値 start を自分で指定したいときは、
この関数の戻り値のリストを得て、
そのリストに対して、初期値の指定が必要な所に初期値の list を与えればよい。
}
\details{
リストの何番目の要素にどういった初期値を指定すればよいかは、
\link[cgd]{nls.freq.all} の出力 cgd の並び順または cgd:::kinds を参照すること。
また、 \link[cgd]{nls.start.template} で、リストの各要素のテンプレートが得られる。
}
\examples{
## preparing
x <- seq( -2, 2, 0.2 )
freq <- c( 15.164, 22.923, 25.134, 27.631, 37.239, 40.464,
		47.126, 79.469, 109.966, 118.241, 111.333, 78.674,
		46.921, 41.026, 36.975, 27.403, 25.493, 22.838,
		14.992, 11.468, 9.174 )
total <- sum( freq )

## getting the initial list
start.list <- init.start.list()

## setting the start parameters of cgd:::kinds[13]
start.list[[13]] <- nls.start.template( 13 )
start.list[[13]]	## checking the names of parameters for setting the start of cgd:::kinds[13]

start.list[[13]]$mean.1.1 <- -0.671
start.list[[13]]$mean.1.2 <- -0.198
start.list[[13]]$mean.2.1 <- 0.293
start.list[[13]]$mean.2.2 <- -0.198
start.list[[13]]$sqrt.sd <- sqrt( 0.640 )	## sqrt.sd is the square root of the standard deviation.

## try nls.freq.all
result <- nls.freq.all( x, freq, total, start = start.list )
result$cor
}
\seealso{
\link[cgd]{nls.freq.all}, \link[cgd]{nls.start.template}
}
