% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{intervals.q.ind}
\alias{intervals.q.ind}
\alias{\S4method{intervals.q.ind}{CGD}}
\title{独立区間の定義域取得}
\usage{
\S4method{intervals.q.ind}{CGD}()
}
\value{
intervals の
         q.ind[1] を並べたベクトルと q.ind[2] を並べたベクトルの data.frame
}
\description{
intervals フィールドの q.ind の値を順に並べた data.frame を取得する。
}
\examples{
 a <- CGD$new()
 a$set.waypoints(
     data.frame( p = c( 0.25, 0.5, 0.75 ), q = c( -0.67, 0, 0.53 ) ),
     this.type1.type = 1 )
 a$intervals.q.ind()

 a$set.waypoints(
     data.frame(
         p = c( 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9 ),
         q = c( -1.40, -0.96, -0.61, -0.30, 0, 0.32, 0.72, 1.23, 2.21 ) ),
 this.type1.type = 2 )
 a$intervals.q.ind()
}
