% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{is.ind.x}
\alias{is.ind.x}
\alias{\S4method{is.ind.x}{CGD}}
\title{独立区間判定 (X座標)}
\usage{
\S4method{is.ind.x}{CGD}(x)
}
\arguments{
\item{x}{X座標のベクトル。}
}
\value{
第1要素 (bool) に、独立区間に入っていれば TRUE、入っていなければ FALSE。
         第2要素 (i) に、当該 intervals のインデックス番号が入ったリスト。
              bool = FALSE の場合は NaN。
}
\description{
指定されたX座標が独立区間に入っているかどうかを調べる。
}
\examples{
 a <- CGD$new()
 a$set.waypoints(
 data.frame(
     p = c( 0.2, 0.5, 0.6, 0.7 ),
     q = c( qnorm( c( 0.2, 0.5, 0.6 ), 0, 1 ), 0.5 ) ),
 this.type1.type = 1 )
 a$is.ind.x( c( -Inf, -0.3, 0.3, 3.0, Inf ) )
}
