% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{mean.conn.bound}
\alias{mean.conn.bound}
\title{[内部関数] 部分期待値算出 (接続区間)}
\usage{
\method{mean}{conn.bound}(
  type1.type,
  itv.1.or.median,
  itv.2.or.sd.1,
  sd.2 = NULL,
  x = c(-Inf, Inf),
  thsd.p.sum = 2e-04,
  force.type1 = !inherits(itv.1.or.median, "CGDInterval")
)
}
\arguments{
\item{type1.type}{接続区間が type 1 の場合の計算方法。
type1.type = 0 の場合、他の引数の値に関係なく、
この関数は常に 0 を返す。}

\item{itv.1.or.median}{接続区間の下側の正規分布の
\link[cgd]{CGDInterval} クラスオブジェクト。
または、分布の中央値 (numeric)。
この値は、各構成要素の正規分布の平均値でもある。}

\item{itv.2.or.sd.1}{接続区間の上側の正規分布の
\link[cgd]{CGDInterval} クラスオブジェクト。
または、下側の正規分布の標準偏差 (numeric)。}

\item{sd.2}{接続区間の上側の正規分布の標準偏差 (numeric)。}

\item{x}{接続区間の下限値 (x[1]) と上限値 (x[2])。
引数 x は、第3～第6引数の型が numeric の場合に用いられる。}

\item{thsd.p.sum}{type1.type = 2 の場合、
2つの正規分布の累積分布関数の値の差の絶対値が、
この値よりも小さい場合は、
期待値を数値積分 (\link[stats]{integrate}) で計算する
 (デフォルト: 2e-4)。
この処置の目的は、累積分布関数の差の絶対値が小さいとき、
\link[stats]{integrate} を使わないで期待値を計算しようとすると、
著しい桁落ち誤差が生じることがあるので、それを防ぐためである。}

\item{force.type1}{TRUEの場合、実際の接続区間の type に関係なく、
強制的に type 1 として計算する。計算テスト用のオプション
 (デフォルト: !inherits( itv.1.or.median, "CGDInterval" ))。}
}
\value{
与えられた接続区間における部分期待値。
}
\description{
不連続分布の平均値 (期待値) の計算のうち、接続区間における部分の計算をする。
type1.type = 1 では、 \link[stats]{integrate} は使用しない。
type1.type = 2 では、 \link[stats]{integrate} を使用しないときに生じる
桁落ち誤差のリスクが、 \link[stats]{integrate} で生じる誤差のリスクを
上回ることがあるので、 \link[stats]{integrate} を使用する場合がある。
type1.type = 0 では、この関数は常に 0 を返す。
}
