% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{mean.cont}
\alias{mean.cont}
\title{[内部関数] 平均値算出 (連続分布)}
\usage{
\method{mean}{cont}(type1.type, means, sds)
}
\arguments{
\item{type1.type}{type1.typeの値。}

\item{means}{intervals の構成要素の正規分布の平均値のベクトル。}

\item{sds}{intervals の構成要素の正規分布の標準偏差のベクトル。}
}
\value{
連続な連結ガウス分布の平均値。
}
\description{
連続な連結ガウス分布の平均値を求める。 \link[stats]{integrate} は使用しない。
本関数は type1.type = 2, symmetric = TRUE の場合に対応していないので、
type1.type = 2, symmetric = TRUE のとき、本関数を使用してはならない。
なお、そのときは mean = median である。
type1.type = 4 の場合、 means と sds の要素数は 4 でも 6 でも同様に計算可能。
}
