% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{mean.nd.bound}
\alias{mean.nd.bound}
\title{[内部関数] 部分期待値算出 (正規分布)}
\usage{
\method{mean}{nd.bound}(obj.or.mean, sd = NULL, x = c(-Inf, Inf))
}
\arguments{
\item{obj.or.mean}{計算対象の独立区間を持つ interval の要素
 (\link[cgd]{CGDInterval} クラスオブジェクト)。
または構成要素の正規分布の平均値 (numeric)。}

\item{sd}{構成要素の正規分布の標準偏差。}

\item{x}{計算対象の閉区間の下限値 (x[1]) と上限値 (x[2])。
引数 sd と x は、第1引数の型が numeric の場合に用いられる。}
}
\value{
与えられた閉区間における正規分布の部分期待値。
}
\description{
ある閉区間 [x[1], x[2]] における、正規分布の部分期待値を算出する。
すなわち、 \eqn{\int_{x_1}^{x_2} t f(t) dt} の計算をする。
\link[stats]{integrate} は使用しない。
}
