% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{p}
\alias{p}
\alias{\S4method{p}{CGD}}
\title{累積分布関数}
\usage{
\S4method{p}{CGD}(q)
}
\arguments{
\item{q}{X座標のベクトル。}
}
\value{
確率変数が q 以下になる確率
}
\description{
X座標を指定して、確率変数がその値以下になる確率を求める。
}
\examples{
 a <- CGD$new()
 a$set.waypoints(
     data.frame( p = c( 0.25, 0.5, 0.75 ), q = c( -0.67, 0, 0.53 ) ),
     this.type1.type = 2, continuous = TRUE )
 a$p( c( -0.67, 0, 0.53 ) )
 plot( seq( -3, 3, 0.01 ), a$p( seq( -3, 3, 0.01 ) ), type = "l" )
}
