% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{q.manage}
\alias{q.manage}
\alias{\S4method{q.manage}{CGDInterval}}
\title{負担区間取得}
\usage{
\S4method{q.manage}{CGDInterval}()
}
\value{
負担区間のX座標
}
\description{
負担区間 (前の区間との接続区間～次の区間との接続区間) に対するX座標を取得する。
}
\examples{
 a <- CGD$new()
 a$intervals[[1]]$q.manage()
}
