% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{q}
\alias{q}
\alias{\S4method{q}{CGD}}
\title{クォンタイル関数}
\usage{
\S4method{q}{CGD}(prob, tol = .Machine$double.eps * 16)
}
\arguments{
\item{prob}{確率のベクトル。}

\item{tol}{許容誤差 (デフォルト: .Machine$double.eps * 16)。}
}
\value{
確率 prob に当たる累積分布関数のX座標
}
\description{
確率を指定して、累積分布関数において対応するX座標を求める。累積分布関数の逆関数。
}
\examples{
 a <- CGD$new()
 a$set.waypoints(
     data.frame( p = c( 0.25, 0.5, 0.75 ), q = c( -0.67, 0, 0.53 ) ),
     this.type1.type = 2, continuous = TRUE )
 a$q( c( 0.25, 0.5, 0.75 ) )
 plot( seq( 0, 1, 0.01 ), a$q( seq( 0, 1, 0.01 ) ), type = "l" )
}
