% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{r}
\alias{r}
\alias{\S4method{r}{CGD}}
\title{乱数生成}
\usage{
\S4method{r}{CGD}(n, tol = 2^(-17))
}
\arguments{
\item{n}{生成する乱数の個数。要素数が 2 以上のベクトルを指定した場合は、
その要素数と同じ個数の乱数を生成する。}

\item{tol}{q() に許容する誤差 (デフォルト: 2^(-17))。}
}
\value{
乱数のベクトル
}
\description{
確率分布に従う乱数を生成する。
}
\examples{
 a <- CGD$new()
 a$set.waypoints(
 data.frame( p = c( 0.25, 0.5, 0.75 ), q = c( -0.67, 0, 0.53 ) ),
 this.type1.type = 2, continuous = TRUE )
 hist( a$r( 400 ) )
}
