% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{separate.freq.four}
\alias{separate.freq.four}
\title{[内部関数] 度数分布データ四分割}
\usage{
separate.freq.four(x, freq, data.mean)
}
\arguments{
\item{x}{X座標のベクトル。
最低 3 個以上の要素を含み、昇順に並んでいて、値の重複はない。}

\item{freq}{度数分布のベクトル。}

\item{data.mean}{データの平均値。}
}
\value{
四分割したX座標と度数分布のベクトルのリスト
}
\description{
度数分布をデータの平均値で二分割し、さらにそれぞれをX座標の個数が等しくなるように二分割する。
ちょうど平均値と等しいX座標がある場合は、2番目と3番目のデータの両方にその点を入れる。
最初に二分割したとき、X座標が奇数の場合は、分割後の前後のデータの両方に中央の点を入れる。
}
