% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{separate.two.four}
\alias{separate.two.four}
\title{[内部関数] 度数分布四分割}
\usage{
separate.two.four(x, freq, data.mean)
}
\arguments{
\item{x}{X座標のベクトル。
最低 3 個以上の要素を含み、昇順に並んでいて、値の重複はない。}

\item{freq}{度数分布のベクトル。}

\item{data.mean}{データの平均値。}
}
\value{
四分割したX座標と度数分布のベクトルのリスト
}
\description{
度数分布をデータの平均値で二分割し、さらにそれぞれをX座標の個数で二分割する。
ちょうど平均値と等しいX座標がある場合は、
リストの2番目と3番目のベクトルの両方にその度数を入れる。
二分割した結果、個数が奇数だった場合は、リストの前後のベクトルの両方に中央の度数を入れる。
}
