% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{t3.mid.sd}
\alias{t3.mid.sd}
\title{[内部関数] 山側の正規分布の標準偏差算出}
\usage{
t3.mid.sd(mean.2, wp.near)
}
\arguments{
\item{mean.2}{山側の正規分布の平均値。}

\item{wp.near}{中央値に近い経路の2点 (以上) の座標。}
}
\value{
山側の正規分布の標準偏差
}
\description{
type1.type=3, 5点経路の場合および、中央値の指定がない3点経路の場合の、
山側 (2番目) の正規分布の標準偏差を計算する。
}
\details{
nleqslv でワーニングが出ないようにするために、この関数では、
wp.near と mean.2 の位置関係がおかしい場合は、内部で勝手にクォンタイルの値を書き換えている。
そのため、この関数は nleqslv で標準偏差を探索する目的以外に用いてはならない。
}
