% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{t3.v.grad.wp4.intervals}
\alias{t3.v.grad.wp4.intervals}
\title{[内部関数] intervals 導出 (4点経路, 2つの正規分布の縦方向グラデーション)}
\usage{
t3.v.grad.wp4.intervals(wp, control)
}
\arguments{
\item{wp}{経路。}

\item{control}{nleqslv に渡す設定。}
}
\value{
intervals, result (nleqslv の結果) のリスト
}
\description{
4点経路の場合に、 2つの正規分布の縦方向グラデーションの intervals を導出する
}
