% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{t3.wp4.intervals}
\alias{t3.wp4.intervals}
\title{[内部関数] type1.type=3, v.grad = FALSE, 4点経路の intervals を導出する}
\usage{
t3.wp4.intervals(wp, uni.sigma, control)
}
\arguments{
\item{wp}{経路。}

\item{uni.sigma}{標準偏差を揃えるかどうかのフラグ。}

\item{control}{nleqslv に渡す設定。}
}
\value{
intervals, result (nleqslv の結果) のリスト
}
\description{
[内部関数] type1.type=3, v.grad = FALSE, 4点経路の intervals を導出する
}
