% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{v.conn.bound}
\alias{v.conn.bound}
\title{[内部関数] 部分分散算出 (接続区間)}
\usage{
v.conn.bound(
  type1.type,
  mean,
  itv.1.or.median,
  itv.2.or.sd.1,
  sd.2 = 0,
  q = list(numeric(), numeric()),
  x = q,
  thsd.p.sum = 2e-04,
  force.type1 = !inherits(itv.1.or.median, "CGDInterval")
)
}
\arguments{
\item{type1.type}{接続区間が type 1 の場合の計算方法。
type1.type = 0 の場合、他の引数の値に関係なく、
この関数は常に 0 を返す。}

\item{mean}{分布全体の平均値。}

\item{itv.1.or.median}{接続区間の下側の正規分布の
\link[cgd]{CGDInterval} クラスオブジェクト。
または、分布の中央値 (numeric)。}

\item{itv.2.or.sd.1}{接続区間の上側の正規分布の
\link[cgd]{CGDInterval} クラスオブジェクト。
または、接続区間の下側の正規分布の標準偏差 (numeric)。}

\item{sd.2}{接続区間の上側の正規分布の標準偏差 (デフォルト: 0)。}

\item{q}{接続区間のX座標の下限値 (q[1]) と上限値 (q[2])
(デフォルト: list( numeric(), numeric() ))。}

\item{x}{計算対象範囲の下限値 (x[1]) と上限値 (x[2])
 (デフォルト: q)。
x の範囲は q の範囲に包含されている必要がある。}

\item{thsd.p.sum}{type1.type = 2 の場合、
2つの正規分布の累積分布関数の値の差の絶対値が、
この値よりも小さい場合は、
期待値を数値積分 (\link[stats]{integrate}) で計算する
 (デフォルト: 2e-4)。
この処置の目的は、累積分布関数の差の絶対値が小さいとき、
\link[stats]{integrate} を使わないで分散を計算しようとすると、
著しい桁落ち誤差が生じることがあるので、それを防ぐためである。}

\item{force.type1}{TRUEの場合、実際の接続区間の type に関係なく、
強制的に type 1 として計算する。計算テスト用のオプション
 (デフォルト: !inherits( itv.1.or.median, "CGDInterval" ))。}
}
\value{
与えられた計算対象範囲における部分分散。
}
\description{
不連続分布の分散の計算のうち、接続区間における部分分散を算出する。
type1.type = 1 では、 \link[stats]{integrate} は使用しない。
type1.type = 2 では、 \link[stats]{integrate} を使用しないときに生じる
桁落ち誤差のリスクが、 \link[stats]{integrate} で生じる誤差のリスクを
上回ることがあるので、 \link[stats]{integrate} を使用する場合がある。
type1.type = 0 では、この関数は常に 0 を返す。
}
