% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{v.cont.t4.via.integrate}
\alias{v.cont.t4.via.integrate}
\title{[内部関数] 数値積分による分散算出 (type1.type = 4)}
\usage{
v.cont.t4.via.integrate(
  means,
  sds,
  mean = mean.cont(4, means, sds),
  get.lv = FALSE,
  get.uv = FALSE
)
}
\arguments{
\item{means}{intervals の構成要素の正規分布の平均値のベクトル。}

\item{sds}{intervals の構成要素の正規分布の標準偏差のベクトル。}

\item{mean}{分布全体の平均値
(デフォルト: mean.cont( 4, means, sds ))。}

\item{get.lv}{TRUE の場合、下半分散を算出する (デフォルト: FALSE)。}

\item{get.uv}{TRUE の場合、上半分散を算出する (デフォルト: FALSE)。
get.lv と get.uv を両方同時に TRUE にすべきではないが、
もし両方とも TRUE の場合は、 get.lv が優先される。
両方とも FALSE の場合は、分布全体の分散を算出する。}
}
\value{
半分散 (または分散) を求めた \link[stats]{integrate} の出力リスト。
}
\description{
type1.type = 4 の場合に、数値積分 (\link[stats]{integrate}) を使って
下半分散または上半分散を求める。分布全体の分散を求めることも可能。
means と sds の要素数は 4 でも 6 でも同様に計算可能。
}
