% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{v.nd.bound}
\alias{v.nd.bound}
\title{[内部関数] 部分分散算出 (正規分布)}
\usage{
v.nd.bound(mean, mean.nd, sd.nd, x)
}
\arguments{
\item{mean}{分布全体の平均値。}

\item{mean.nd}{正規分布の平均値。}

\item{sd.nd}{正規分布の標準偏差。}

\item{x}{計算対象の閉区間の下限値 (x[1]) と上限値 (x[2])
(デフォルト: c( -Inf, Inf ))。}
}
\value{
与えられた閉区間における部分分散。
}
\description{
ある閉区間 [x[1], x[2]] において、確率分布が (局所的に) 正規分布に従うとき、
その閉区間における部分分散を算出する。
すなわち、 \eqn{\int_{x_1}^{x_2} (t - \mu)^2 f(t) dt} の計算をする。
ここで、 \eqn{\mu} は分布全体の平均値である。
\link[stats]{integrate} は使用しない。
}
