% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{v.sub}
\alias{v.sub}
\title{[内部関数] 分散計算用サブ関数 (type1.type = 1, 2, 3)}
\usage{
v.sub(type1.type, mean, mean.i, sd.i, x, q = x, p.sum = 0, k = 0)
}
\arguments{
\item{type1.type}{接続区間が type 1 の場合の計算方法}

\item{mean}{分布全体の平均値。}

\item{mean.i}{intervals の構成要素の正規分布の平均値。}

\item{sd.i}{intervals の構成要素の正規分布の標準偏差。}

\item{x}{分散の計算対象範囲の下限値または上限値。}

\item{q}{接続区間のX座標の下限値または上限値 (デフォルト: x)。
type1.type = 2 の連続分布の場合は
下限値として -Inf 、上限値として Inf を与えること。}

\item{p.sum}{type1.type = 2 の場合に使用する。
2つの正規分布の累積分布関数の合計値 (デフォルト: 0)。}

\item{k}{type1.type = 3 の場合に使用する。
intervals を構成する3つの正規分布のうち、
何番目について計算するかを指定する (デフォルト: 0)。}
}
\value{
分散計算用サブ関数の値。
}
\description{
連続分布や不連続分布の接続区間の部分の分散の計算に用いるサブ関数。
サブ関数の意味や表現式は、 type1.type の値によって、それぞれ異なる。
\link[stats]{integrate} は使用しない。
}
\details{
この関数では、 type1.type の値によって、
 それぞれ以下の計算を \link[stats]{integrate} を使わずに計算する。
 最終的な分散の値はこれらの計算値の和と差および、簡単な項との和や積で表される。

 以下の式において、 \eqn{\mu} は分布全体の平均値 (mean)、
 \eqn{f_i(x)} は平均値 mean.i 、標準偏差 sd.i の正規分布の確率密度関数、
 同じく \eqn{\Phi_i(x)} は累積分布関数、
 \eqn{\bar \Phi(x)} は2つの累積分布関数 \eqn{\Phi_1(x)} と \eqn{\Phi_2(x)} の平均である。

 \describe{
     \item{type1.type = 1}{
         \deqn{3 \left \lbrace (x - \mu)^2 (q - x) \Phi_i(x) -
               2 \displaystyle
                 \int_{-\infty}^x (t - \mu) (q - t) \Phi_i(t) \ dt \right \rbrace}}

     \item{type1.type = 2}{
         \deqn{\displaystyle
               \int_{-\infty}^x (t - \mu)^2 (\Phi_i(t) - \bar \Phi(q)) f_i(t) \ dt}}

     \item{type1.type = 3}{
         \deqn{k = 1, 3 \ :
             \ \displaystyle
               \int_{-\infty}^x (t - \mu) (1 - \dfrac{f_i(t)}{f_i(\mu_i)}) f_i(t) \ dt}

         \deqn{k = 2 \ :
             \ \displaystyle
               \int_{-\infty}^x (t - \mu) \dfrac{f_i(t)^2}{f_i(\mu_i)} f_i(t) \ dt}}
 }
}
