% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{variance.sub}
\alias{variance.sub}
\title{正規分布の分散の閉区間範囲内の値を返す}
\usage{
variance.sub(
  x.bound,
  mean,
  sd,
  f.d = function(x, mean, sd) dnorm(x, mean, sd),
  f.p = function(x, mean, sd) pnorm(x, mean, sd)
)
}
\arguments{
\item{x.bound}{閉区間を指定する、要素2個のベクトル (x.bound[1] < x.bound[2] であること)}

\item{mean}{正規分布の平均}

\item{sd}{正規分布の標準偏差}

\item{f.d}{正規分布の確率密度を計算する関数 (デフォルト: dnorm( x, mean, sd ) )}

\item{f.p}{正規分布の確率を計算する関数 (デフォルト: pnorm( x, mean, sd ) )}
}
\value{
与えられた正規分布の分散を閉区間範囲内のみに限って計算した値
}
\description{
正規分布の分散の閉区間範囲内の値を返す
}
