% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\docType{class}
\name{CGD-class}
\alias{CGD-class}
\alias{CGD}
\title{連結ガウス分布 (Connected Gaussian Distribution) クラス}
\description{
複数の正規分布をX軸方向に連結したモデルを表すクラス。
}
\details{
連結ガウス分布は、
理論上、任意のクォンタイルをゼロ誤差で満たす分布を構成できる。
そのため、何かクォンタイルが与えられたときに、
それを高精度に再現する分布の概形やサンプルが欲しいときなどに使うことができる。
ただし、確率密度関数は一般に不連続であり、歪んだ形になる。

単峰性や二峰性の連続分布を構成することも可能であるが、
連続分布を構成するには、クォンタイルの個数が8点以下で、
クォンタイルの位置が過度にいびつでないことが条件になる。

不連続な連結ガウス分布を使って滑らかなグラフを得るには、
ランダムサンプルを取得して、そのヒストグラムを描画するとよい。

また、 \link[cgd]{nls.freq} や \link[cgd]{nls.freq.all} を使うと、
データの度数分布をもとに、その分布に近い連続分布モデルを得ることもできる。
このとき、データは正規分布に従っていなくてもよい。
むしろ、このクラスはデータが正規分布に従わない場合
 (あるいは、正規分布に従う裏付けが無い場合) に用いるべきである。
}
\section{Fields}{

\describe{
\item{\code{kind.index}}{この分布の種類のインデックス番号。}

\item{\code{kind}}{この分布の種類名。}

\item{\code{median}}{この分布の中央値。}

\item{\code{intervals}}{連結区間 (\link[cgd]{CGDInterval} クラスのリスト)。}

\item{\code{type1.type}}{接続区間が type 1 の場合の計算方法
(詳細は \link[cgd]{trace.q} の Details を参照)。}

\item{\code{mean}}{この分布の平均値。}

\item{\code{sd}}{この分布の標準偏差。}

\item{\code{lsd}}{この分布の下半標準偏差。}

\item{\code{usd}}{この分布の上半標準偏差。}

\item{\code{lsd.abs.error}}{下半標準偏差の絶対誤差の推定値。}

\item{\code{usd.abs.error}}{上半標準偏差の絶対誤差の推定値。
type1.type = 4 の場合、
半標準偏差の計算に \link[stats]{integrate} を用いるが、
それに由来する絶対誤差がこれらのフィールドに入る。}
}}


\seealso{
\link[cgd]{CGDInterval}, \link[cgd]{set.waypoints}, \link[cgd]{trace.q},
         \link[cgd]{nls.freq}, \link[cgd]{nls.freq.all},
         \href{https://github.com/Kimitsuna-Goblin/CGD}{README.md} (GitHub)
}
