% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\docType{class}
\name{CGD-class}
\alias{CGD-class}
\alias{CGD}
\title{連結ガウス分布 (Connected Gaussian Distribution) クラス}
\description{
平均値の等しい正規分布をX軸方向に連結したモデルを表すクラス
}
\details{
連結ガウス分布は、任意のクォンタイルをゼロ誤差で満たす分布を構成できる。
そのため、何かクォンタイルが与えられたときに、
それを高精度に再現する分布が欲しいときなどに使うことができる。
ただし、確率密度関数は一般に不連続であり、歪んだ形になる。

クォンタイルが8点以下の場合は、
type1.type の設定を適切に与えることにより、
クォンタイルの位置が過度にいびつでなければ、
確率密度関数が連続で滑らかな分布を構成することも可能である。

確率密度関数が不連続な連結ガウス分布を使って滑らかなグラフを得るには、
ランダムサンプルを取得して、そのヒストグラムを描画するとよい。
}
\section{Fields}{

\describe{
\item{\code{mean}}{平均値}

\item{\code{intervals}}{連結区間 (\link[cgd]{CGDInterval-class} クラスのリスト)}

\item{\code{type1.type}}{接続区間が type 1 の場合の計算方法 (詳細は \link[cgd]{CGD$set.waypoints} を参照)}

\item{\code{m.sd}}{計算済みの標準偏差 (クラス外からは直接参照しないこと)}

\item{\code{m.lsd}}{計算済みの下側標準偏差 (クラス外からは直接参照しないこと)}

\item{\code{m.usd}}{計算済みの上側標準偏差 (クラス外からは直接参照しないこと)}
}}


\seealso{
\link[cgd]{CGDInterval-class}
\link[cgd]{CGD$set.waypoints}
\href{https://github.com/Kimitsuna-Goblin/CGD}{README.md} (GitHub)
}
