% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$initialize}
\alias{CGD$initialize}
\title{コンストラクタ}
\usage{
a <- CGD$new(mean = NULL, intervals = NULL, type1.type = 1)
}
\arguments{
\item{mean}{平均値 (デフォルト: NULL)}

\item{intervals}{連結区間 (CGDInterval クラスのリスト) (デフォルト: NULL)}

\item{type1.type}{type 1 の場合の計算方法 (1、2、3 のいずれかを指定する) (デフォルト: 1)}
}
\description{
コンストラクタの引数は予め連結区間の構成が分かっている場合に指定する。
詳細については、 set.waypoints() を参照。
}
\seealso{
\link[cgd]{CGDInterval-class}
\link[cgd]{CGD$set.waypoints}
}
