% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$r}
\alias{CGD$r}
\title{ランダムサンプル取得}
\usage{
a <- CGD$new()
a$r(n, tol = 2^(-17))
}
\arguments{
\item{n}{サンプル数}

\item{tol}{q() に許容する誤差 (デフォルト: 2^(-17))}
}
\value{
ランダムサンプルのベクトル
}
\description{
ランダムサンプルを取得する
}
