% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$set.intervals}
\alias{CGD$set.intervals}
\title{連結区間設定}
\usage{
a <- CGD$new(type1.type = 1)
a$set.intervals(intervals, this.type1.type = NULL)
}
\arguments{
\item{intervals}{連結区間 (\link[cgd]{CGDInterval-class} クラスのリスト)}

\item{this.type1.type}{フィールドの type1.type に設定する値 (1、2、3、4 のいずれかを指定すること)。
NULL の場合は type1.type の値を変更しない (デフォルト: NULL)}
}
\description{
予め連結区間の構成が分かっている場合に、連結区間を設定する。
フィールド mean は intervals の内容に応じて適切に設定され、
フィールド m.sd, m.lsd, m.usd は初期化される。
詳細については、 \link[cgd]{CGD$set.waypoints} を参照。
}
\seealso{
\link[cgd]{CGDInterval-class} \link[cgd]{CGD$set.waypoints}
}
