% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{CGD$set.waypoints}
\alias{CGD$set.waypoints}
\title{経路設定}
\usage{
a <- CGD$new()
a$set.waypoints(waypoints, continuous = FALSE, symmetric = FALSE, uni.sigma = FALSE, this.type1.type = NULL)
}
\arguments{
\item{waypoints}{経路の data.frame( q = 経路のX座標 (クォンタイル), p = その点における確率 )
X座標 (クォンタイル) は昇順にソートしておくこと。
平均値は p = 0.5 の点のX座標として与えること
 (type1.type = 2 で continuous = TRUE、または、type1.type = 3 の場合を除き、p = 0.5 の点は必須)}

\item{continuous}{独立区間を [0, 0] と [1, 1] の2点にして、
確率密度関数が全区間 (-∞, ∞) で連続になるように分布構成を試みる。
type1.type = 1, 2 で有効 (type1.type >= 3 では常に連続) (デフォルト: FALSE)}

\item{symmetric}{独立区間を [0, 0], [0.5, 0.5], [1, 1] の3点にして、
1番目と3番目の確率分布を同一にすることにより、
確率密度関数が全区間 (-∞, ∞) で連続で、かつ左右対称になるように試みる。
type1.type = 1, 2, 3 で有効 (デフォルト: FALSE)}

\item{uni.sigma}{type1.type = 2、continuous = TRUE で、かつ、経路の構成点が3点の場合に、
2つの確率密度関数 f_1(x) と f_2(x) の標準偏差を同じ値に揃えるかどうかのフラグ。
このフラグが TRUE のとき、通常、 f_1(x) と f_2(x) の平均値は等しくならない (デフォルト: FALSE)}

\item{this.type1.type}{フィールドの type1.type に設定する値 (1、2、3 のいずれかを指定すること)。
NULL の場合は type1.type の値を変更しない (デフォルト: NULL)

type1.type = 1 は最も単純な連結である。 continuous = TRUE にすると、2つの正規分布の平均が構成可能である。

type1.type = 2 は continuous = TRUE にすると、2つの確率密度関数の横方向グラデーションが構成可能である。

type1.type = 3 は symmetric = TRUE にすると、2つの確率密度関数の縦方向グラデーションが構成可能である。

type1.type の値によって、接続区間 (β_i, α_{i+1}) が type 1 の場合、
接続区間の累積分布関数 Ψ_i(x) を以下のように計算する。

1: Ψ_i(x) = ( α_{i+1} - x ) / ( α_{i+1} - β_i ) * Φ_i(x) +
             ( x - β_i ) / ( α_{i+1} - β_i ) * Φ_{i+1}(x)

2: Ψ_i(x) = ( Φ~_i(α_{i+1}) - Φ~_i(x) ) / ( Φ~_i(α_{i+1}) - Φ~_i(β_i) ) * Φ_i(x) +
             ( Φ~_i(x) - Φ~_i(β_i) ) / ( Φ~_i(α_{i+1}) - Φ~_i(β_i) ) * Φ_{i+1}(x)

3: Ψ(x) = ∫_{-∞}^{min( x, μ_1 )} ( 1 - f_1(t) / f_1(μ_1) ) f_1(t) dt
            + ∫_{-∞}^x f_2(t)^2 / f_2(μ) dt
            + ∫_{min( x, μ_3 )}^x ( 1 - f_3(t) / f_3(μ_3) ) f_3(t) dt
         = min( Φ_1(x) - Φ^*_1(x) / √2, ( 2 - √2 ) / 4 )
            + Φ^*_2(x) / √2
            + max( 0, Φ_3(x) - Φ^*_3(x) / √2 - ( 2 - √2 ) / 4 )

ただし、Φ_i, Φ_{i+1} は当該接続区間の前後の独立区間を負担する正規分布の累積分布関数。
Φ~_i(x) = ( Φ_i(x) + Φ_{i+1}(x) ) / 2。
f_i, f_{i+1} は当該接続区間の前後の独立区間を負担する正規分布の確率密度関数。μ は平均値。
Φ^*_i は正規分布 N(μ_i, (σ_i / √2)^2) の累積分布関数。

type1.type = 1 では、
set.waypoints() の引数で continuous または symmetric を TRUE にした場合 (2つの正規分布の平均)、
独立区間は [0, 0], [1, 1] の2点になり、
接続区間の累積分布関数は Ψ(x) = ( Φ_i(x) + Φ_{i+1}(x) ) / 2 となる。

type1.type = 2 では、
set.waypoints() の引数で continuous を TRUE にした場合 (横方向グラデーション)、
独立区間は [0, 0], [1, 1] の2点になり、
接続区間の累積分布関数は上の 2 の式になる。
set.waypoints() の引数で symmetric を TRUE にした場合、
独立区間は [0, 0], [0.5, 0.5], [1, 1] の3点になり、
接続区間 (0, 0.5) の累積分布関数 Ψ_1(x) は上の 2 の式で、
接続区間 (0.5, 1) の累積分布関数 Ψ_2(x) は Ψ_2(x) = 1 - Ψ_1(2μ - x) となる。

type1.type = 3 では、
独立区間は [0, 0], [0.5, 0.5], [1, 1], [0, 0.5], [0.5, 1], [0, 1] の6種類しか取り得ない。
set.waypoints() の引数の continuous は無効になる。
set.waypoints() の引数で symmetric を TRUE にした場合 (縦方向グラデーション)、
独立区間は [0, 0], [0.5, 0.5], [1, 1] の3点になり、
接続区間 (0, 0.5), (0.5, 1) の累積分布関数はどちらも同じ、上の 3 の式になる。

continuous = TRUE or symmetric = TRUE or type1.type = 3 の場合、経路の構成点は一定の個数でなければならない。
具体的には、構成点の個数は以下のようにする必要がある (個数に (*) が付いているものは、確率 0.5 の点が必須)。

・continuous = TRUE :   type1.type = 1 ⇒ 3点(*)、
                        type1.type = 2 ⇒ 3点(*) or 4点

・symmetric = TRUE :    type1.type = 1 or 2 ⇒ 3点(*)

・type1.type = 3 :      symmetric = FALSE ⇒ 3点(*) or 4点 or 5点(*) or 6点、
                        symmetric = TRUE ⇒ 3点(*)

個数に (*) が付いていないものは、構成点の半数が確率 0.5 未満、半数が 0.5 超であることが望ましい。
ただし、type1.type = 3 で 4点経路の場合は、構成点がどちらか一方に偏っていてもよい。}
}
\value{
nleqslv() を内部で実行した場合はその結果。それ以外は NULL
}
\description{
累積分布関数の経路 (クォンタイル) を設定する
}
\seealso{
\href{https://github.com/Kimitsuna-Goblin/CGD}{README.md} (GitHub)
}
