% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{ms.qp.norm}
\alias{ms.qp.norm}
\title{正規分布の平均値・標準偏差算出}
\usage{
ms.qp.norm(q, p)
}
\arguments{
\item{q}{X座標 (クォンタイル)。要素数 2 個のベクトル。}

\item{p}{それらのX座標における確率 (累積分布関数の値)。要素数 2 個のベクトル。}
}
\value{
list( mean = 平均値, sd = 標準偏差 )
}
\description{
与えられた2点のクォンタイルと確率を取るような累積分布関数を持つ正規分布の、平均値と標準偏差を得る。
}
\examples{
ms.qp.norm( q = c( -1, 1 ), p = pnorm( c( -1, 1 ), 0, 1 ) ) # list( mean = 0, sd = 1 )
ms.qp.norm( q = c(	0, 1 ), p = pnorm( c(  0, 1 ), 0, 2 ) ) # list( mean = 0, sd = 2 )
ms.qp.norm( q = c( -2, 1 ), p = c( 0.3, 0.7 ) ) # list( mean = 0.5, sd = 2.86 ) (about)
}
