% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGD.R
\name{sd.mqp.norm}
\alias{sd.mqp.norm}
\title{正規分布の標準偏差算出}
\usage{
sd.mqp.norm(mean, q, p)
}
\arguments{
\item{mean}{平均値。}

\item{q}{X座標 (クォンタイル)。平均値と同じ値を与えてはならない。}

\item{p}{そのX座標における確率 (累積分布関数の値)。 0.5 を与えてはならない。}
}
\value{
標準偏差
}
\description{
与えられた平均値を取り、与えられた1点のクォンタイルにおける確率を満たす正規分布の、
標準偏差を得る。
}
\examples{
 sd.mqp.norm( 0, qnorm( 0.3, 0, 1 ), 0.3 ) # 1
 sd.mqp.norm( rep( 0, 5 ), 1:5, pnorm( 1:5, 0, 1 ) ) # 1 1 1 1 1
 sd.mqp.norm( c( -0.1, 0, 0.3 ), c( -0.3, -0.1, 0.4 ), c( 0.38, 0.47, 0.53 ) ) # [2] == [3]
}
